Update Azure Key Vault for Disk Encryption incl applying disk encryption to VM disk
Added Operation Insights workspace for audit logging on the key vault.
Note: This is an untested version, as I seem to be unable to get a working parameter (-workspaceid) used during executing.
Where I have seen someone else being able to use this parameter, however being unable to input it with a correct ID as value for it.

Anyways, I'm uploading it. Because I believe the code to be correct at this moment in time.